refactor: reorganize project structure and update dependencies
- Updated package.json to simplify linting script and add new dependencies.
- Refactored lazy loading imports in App.tsx to reflect new directory structure.
- Adjusted AdCarousel component to handle Swiper navigation with type safety.
- Modified main.tsx to include React import.
- Moved AffiliateAccount and AffiliateReferral interfaces to types.ts for better organization.
- Deleted unused App.tsx, Login.tsx, and Register.tsx files, consolidating authentication pages into a new auth directory.
- Updated Terms component text for consistency.
- Added dependabot configuration for automatic dependency updates.
